requirements.txt
pdfplumber>=0.9.0
openpyxl>=3.1.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
PyMuPDF>=1.23.0
python-docx>=0.8.11